+++ doc/changelog.rst
@@ -8,6 +8,7 @@ What's New
 * Fixes URLs ignored (if any) on command line when ``--input-file`` is specified.
 * Fixes crash when redirected to a URL that is not HTTP.
 * Implements ``--local-encoding``, ``--remote-encoding``, ``--no-iri``.
+* Implements ``--https-only``.
 * PhantomJS:
 
   * Implements "smart scrolling" that avoids unnecessary scrolling.